Added customizable assets
Added new class - KBImageSource and an associated view controller - to
make it simple to get new images from a directory.  Its prefix is "KB"
because I hope to be able to separate this off as an open-source
library to help other apps do this during development.
Added AFNetworking to the app to simplify the networking code in
KBImageSource.